Put BigQuery-specific AvroUtils in with BigQueryUtils to avoid confusion
